ClusterReport(Consense)	R Documentation

Cluster Report

Description

After the result set has been returned by CompareClustering, generates all tables and relevant figures for that result set.

Usage


ClusterReport(clusterresults, write=FALSE, filebase)

Arguments

clusterresults 	 A result set from CompareClustering.

write

filebase 	 A base file name. Files will be saved in the current directory.

Value

The within metrics are displayed as a table and the cluster results are saved as a graph.
Files are saved in the current working directory. Currently the following files are saved:

filebase within.txt 	 A tab delimited text file of the various metrics, as described in withinmetrics().

filebase clusters.txt 	 A tab-delimted text file of all clusterings by assignment.

filebase jaccard.jpg 	 A Jpeg of the metaclustering dendrogram for the Jaccard Metric, comparing all methods. Note that the clustering is actually done on (1-Jaccard).

filebase jaccard.txt 	 A text file of the actual Jaccard metrics in tab-delimited format.

filebase variation.jpg 	 A Jpeg of the metaclustering dendrogram for the Variation of Information metric, comparing all methods.

filebase variation.txt 	 A text file of the actual Variation metrics in tab-delimited format.

Author(s)

Ted Laderas (laderast@ohsu.edu)

See Also

withinmetrics, betweenmetrics, CompareClustering

Examples


    data(choresults)
    #display results and save files to current working directory
    ClusterReport(choresults, write=TRUE, filebase="choresults")

[Package Consense version 0.1.1 Index]