autoCompareClustering(Consense)	R Documentation

Compare Clusterings Automatically

Description

Given a number of clusters k as input, cluster a dataset using multiple methods.

Usage

autoCompareClustering(data, kinput)

Arguments

data 	 A matrix of values to cluster by row. To cluster by column, simply transpose the matrix. However, note that the row names must be unique, as the framework uses the row names in calculating the between-method metrics.

kinput 	 A desired geometry (number of clusters) k, greater than 1 and less than n (number of genes or samples).

Details

This is a wrapper function for CompareClustering that has suitable defaults for automatically clustering and producing results for a given k (number of clusters). In the case of hierarchical methods, the dendrograms will be pruned at a height that yields that number of clusters.

Note that for SOM, we are essentially limiting it to a one dimensional map (k x 1). If two dimensional maps are required, CompareClustering is a more appropriate function.

Value

resultset 	 A result set that can be read by ClusterReport or \line{IndexGene}.

Author(s)

Ted Laderas (laderast@ohsu.edu)

See Also

link{CompareClusterings}

Examples


  ##load data
  data(chocellcycle)
  ##look for 5 clusters
  resultset <- autoCompareClustering(chocellcycle, kinput=5)
  ##examine results
  ClusterReport(resultset, "test")

[Package Consense version 0.1.1 Index]