betweenmetrics(Consense)	R Documentation

Between-Method Metrics

Description

Given a list of clusterings, calculates the between method metrics between them.

Usage

betweenmetrics(clusterlist)

Arguments

clusterlist 	 a list of lists which contains clusterings and parameters as supplied by clusterList.

Details

Calculates the between-method metrics between all clusterings as supplied by clusterlist.

Value

A list of the following matrices:

jaccardresult 	 a matrix that contains all possible comparisons between clusterings using the jaccard index. currently this is formatted much like a distance matrix, with the columns corresponding to the first 1:n-1 methods and the rows corresponding to the 2:n methods.

variationresult 	 a matrix that contains all possible comparison between clusterings using the variation of information. Similarly formatted to the matrix for jaccardresult.

Author(s)

Ted Laderas (laderast@ohsu.edu)

See Also

clusterList variationInformation jaccard

[Package Consense version 0.1 Index]