choresults(Consense)	R Documentation

Consense Result Set for Cho Cell Cycle

Description

This is a Consense Result set for the Cho Cell Cycle dataset, made available to demonstrate some of the post-clustering functionality.

Usage

data(choresults)

Format

A result set, to be utilized by ClusterReport or IndexGene.

Source

http://nhlbi.ohsu.edu/consense/

References

Yeung, K.Y., D.R. Haynor, and W.L. Ruzzo, Validating clustering for gene expression data. Bioinformatics, 2001. 17:309-18.

Cho, R.J., et al., A genome-wide transcriptional analysis of the mitotic cell cycle. Mol Cell, 1998. 2:65-73

[Package Consense version 0.1 Index]