clusterList(Consense)	R Documentation

Cluster List

Description

Takes a list of clustering methods and parameters and produces the clusterings.

Usage

clusterList(data, methodlist)

Arguments

data 	 A data matrix to be clustered. Note that if there are missing values, how the missing values are handled (interpolation, etc.) is dependent on the clustering method. For an example, look at hclust().

methodlist 	 A list of methods and their associated parameter. Note that the cluster output functions are passed into clusterList, which does the actual clustering.

Details

see ClusterOutput for more details about writing wrapper functions to get your clustering method to work with this function.

Value

A cluster list. A list of lists, each slot containing the following:

clustering 	 A one column data frame whose first column corresponds to the cluster assignments. The row names and ordering remain the same.

params 	 a list of parameters provided to that method. currently provided by getClusterInfo().

Author(s)

Ted Laderas (laderast@ohsu.edu)

See Also

getClusterInfo methodlistAuto

[Package Consense version 0.1 Index]