kmeansOutput(Consense)	R Documentation

Clustering Output functions

Description

Wrapper functions for clustering methods that pass on appropriate parameters and produce standardized output for betweenmetrics and withinmetrics.

Usage


  kmeansOutput(matrix, params)
  dianaOutput(matrix, params)
  agglomOutput(matrix, params)
  somOutput(matrix, params)

Arguments

matrix 	 A matrix of expression values. Commonly, this corresponds to the columns corresponding to samples and the rows corresponding to genes.

params 	 A list of params supplied to the function.

Details

In order to add a clustering method, you must add a wrapper function that implements the interface clusterOutput(data, params). Params can be passed onto the function by rewriting methodlistAuto (the easy way) or by producing a necessary param list by adding code to getClusterInfo (the more difficult way). The wrapper function must translate the output format of the clustering method into the clustering output format described above.

Value

clustering 	 A data frame with only one cluster whose row values are factors that correspond to the cluster that gene belongs to. Note that these must be factors; otherwise the clustering cannot be utilized by the rest of the Consense package

Author(s)

Ted Laderas (laderast@ohsu.edu

[Package Consense version 0.1 Index]