countcluster(Consense)	R Documentation

produce a vector of cluster counts of each cluster

Description

counts the number within each cluster and produces a vector count. Internal function.

Usage

countcluster(output)

Arguments

output 	 a clustering result as supplied by clusterList.

Value

count 	 A vector of counts for each cluster.

[Package Consense version 0.1.1 Index]