getClusterInfo(Consense)	R Documentation

Get Clustering Info

Description

Query the user about appropriate parameters for clustering.

Usage


getClusterInfo(useGUI = FALSE)

Arguments

useGUI 	 Currently not implemented. This is based on TkWidgets code and is too buggy. Leave as false.

Details

This program will query user as to which methods to use on the command line. If hierarchical methods are used, the program will ask the user to either input a desired number of clusters or if they want the option of handcutting their clusters.

Value

methodlist 	 a list of methods and associated parameters passed on to link{clusterList}.

Author(s)

Ted Laderas (laderast@ohsu.edu)

See Also

methodlistAuto

[Package Consense version 0.1 Index]