homogeneity(Consense)	R Documentation

Calculates the homogeneity of a clustering.

Description

Given cluster output from a suitable clustering function, calculate the Homogeneity (from Shamir, et al.) of the clusterings.

Homogeneity is a measure of the within-cluster variance. See Shamir, et al, for how it is calculated.

Usage

homogeneity(clustering, centers, distfunc="euclidean")

Arguments

clustering 	 Cluster output from a clustering function. the output should be in matrix format with the cluster assignments being the first column in that matrix, with the actual data in remaining columns.

centers 	 Cluster Centers as returned by ClusterCenters.

distfunc 	 Distance function to be used. Currently only "euclidean" or "correlation" are supported.

Value

homogeneity 	 An unbounded, non-negative scalar that represents the homogeneity of the clustering.

Author(s)

Ted Laderas (laderast@ohsu.edu)

References

Shamir, et al. Algorithmic Approaches to Clustering Gene Expression Data. in Current Topics in Computational Molecular Biology. MIT Press: Boston.

Examples


    ##calculate homogeneity for Cho et al's clusters
    data(chocellcycle)
    data(choclusters)
    cho <- cbind(as.numeric(choclusters), chocellcycle)
    cent <- clusterCenters(as.data.frame(cho))
    homogeneity(cho, cent)
  

[Package Consense version 0.1 Index]