inspectTree(Consense)	R Documentation

Inspect Dendrograms

Description

Inspect a clustering dendrogram and cut it into a clustering.

Usage

inspectTree(tree, id)

Arguments

tree 	 A dendrogram as produced by hclust or as.hclust

id 	 An identifier to be displayed on the plot.

Details

This is an internal function to the Consense package that utilizes identify to plot and prune clusterings by selecting branch points. When the user is finished, the program transforms the clusterings into the format required by Consense.

Value

clustering

[Package Consense version 0.1 Index]