separation(Consense)	R Documentation

Average Separation of a clustering.

Description

Given cluster output from a suitable clustering function, calculate the separation (from Shamir, et al.) of the clusterings. Homogeneity can be thought of as a measure of the average between-cluster distance within a clustering.

Usage

separation(clustering, centers)

Arguments

clustering 	 Cluster output from a clustering function. the output should be in matrix format with the cluster assignments being the first column in that matrix, with the actual data in remaining columns.

centers 	 Cluster centers from clusterCenters.

Value

separation 	 An unbounded, non-negative scalar that represents the separation of the clustering.

Author(s)

Ted Laderas (laderast@ohsu.edu)

References

Shamir, et al. Algorithmic Approaches to Clustering Gene Expression Data. in Current Topics in Computational Molecular Biology. MIT Press: Boston.

Examples

##=don't run
    ##calculate separation for Cho et al's clusters
    data(chocellcycle)
    data(choclusters)
    cho <- cbind(as.numeric(choclusters), chocellcycle)
    cent <- clusterCenters(as.data.frame(cho))
    separation(cho, cent)
  

[Package Consense version 0.1 Index]