withinmetrics(Consense)	R Documentation

Within-Method Metrics

Description

Framework for calculating the within-method metrics for a list of methods.

Usage

withinmetrics(clusterlist, data)

Arguments

clusterlist 	 a list of lists which contains clusterings and parameters as supplied by clusterList().

data 	 A matrix of expression values.

Details

Given a list of parameters and clusterings as supplied by clusterList, calculates and returns a matrix with all of the relevant within-method clustering metrics.

Currently, withinresult calculates number of clusters, homogeneity, separation and average silhouette width. Note that this average silhouette width is correlation-based. See silhouette() in the cluster package for more information.

Value

withinresult 	 a matrix whose columns correspond to the metrics calculated and whose rows correspond to the methods these methods were calculated on.

Author(s)

Ted Laderas (laderast@ohsu.edu)

See Also

clusterList homogeneity separation silhouette

[Package Consense version 0.1 Index]